upd qpoisson code
repaire spot bug

+++ opencv_qt_demo/matrix.h
@@ -20,7 +20,7 @@ public:
 	SMat1f Laplace(int, QPoint *);  
 	MatrixX4f Getb(QImage, Mat1i, Mat1b, QPoint, QPoint *);
 	Mat1i divGArray(Mat4b, QPoint *, int);
-	Mat4b Poisson(QImage, MatrixX4f, QPoint, QPoint *);
+	Mat4b Poisson(QImage, MatrixX4i, QPoint, QPoint *);
 
 	QPixmap Filtered(QImage, QPixmap, QRect, QMat *);
 	QImage toImg(QImage, SMat1f, Mat1i, Mat1b, QPoint *, QPoint);
@@ -32,7 +32,7 @@ private:
 	Mat1b toEig(Mat4b, QRect);
 	Mat1b cutMat(Mat1b, QRect);
 	Mat4b cutImg(QImage, Mat1b, QRect);
-	MatrixX4f pSolver(SMat1f, MatrixX4f);
+	MatrixX4i pSolver(SMat1f, MatrixX4f);
 
 	int divG(Mat4b, QPoint, int);
 	int Neighbor(Mat4b, Mat1b, QPoint, QPoint, int);
